Can now grab telescope and place in designated zones
A work-in-progress for the telescope grabbing. I created two snapdrop zones for the telescope and made the telescope only be droppable within those two zones, after it has been grabbed.

The next step is to use the VRTK_AutoGrabObject in combination with placing the top-cover of the telescope. What I then want to happen is for the AutoGrabObject script to become activated leading the telescope to be automatically grabbed.

If the telescope is placed in the astronomy setup we want to detect that and then enable the starmap sprite

If the telescope is placed in the experiment setup we want to unveil the hidden work-bench telescope & only grab the top-cover (such that it can be placed somewhere else, beside the telescope)